# api/openapi.yaml
openapi: 3.0.0
info:
  title: NoodleMap API
  version: 0.0.0
...
paths:
  /api/v1/auth/google:
    post:
      operationId: googleAuth
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
  /api/v1/auth/logout:
    post:
      operationId: logout
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
components:
  schemas:
...
    AuthResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'
...
    ErrorBody:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    GoogleAuthRequest:
      type: object
      required:
        - credential
      properties:
        credential:
          type: string
...
    User:
      type: object
      required:
        - id
        - email
        - isAdmin
      properties:
        id:
          $ref: '#/components/schemas/uuid'
        email:
          type: string
        isAdmin:
          type: boolean
    uuid:
      type: string
      description: UUID serialized as a string.
servers:
  - url: /
    description: Current origin
    variables: {}